bigdata/requirements.txt
asgiref==3.5.0
Django==3.2.12
djangorestframework==3.13.1
joblib==1.1.0
mysqlclient==2.1.0
numpy==1.22.3
pandas==1.4.1
python-dateutil==2.8.2
pytz==2022.1
scikit-learn==1.0.2
scipy==1.8.0
six==1.16.0
sqlparse==0.4.2
threadpoolctl==3.1.0
